Refactor the 'init' command out of lib.rs.
This commit refactors the `init` command implementation out of lib.rs into a
command implementation.
