+++ azure-functions/src/commands.rs
@@ -1,3 +1,5 @@
+mod init;
 mod run;
 
+pub use self::init::*;
 pub use self::run::*;
